### 2023-05-04 #5437 @editorialbot
[PRE REVIEW]: otoole: OSeMOSYS Tools for Energy Work
**Submitting author:** @trevorb1 (<a href="http://orcid.org/0000-0003-2458-2968">Trevor Barnes</a>)
**Repository:** https://github.com/OSeMOSYS/otoole
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.3
