### 2023-05-06 #5437 @trevorb1
Hi @fraukewiese. Looking through the list provided by JOSS, suitable reviewers that we have no conflicts of interest with may be:
- brynpickering
- fneum
- olejandro
- zoltanmaric

If you need more names, please just let me know!
