### 2023-06-20 2729b45
Adds a public API to convert and convert_results methods
In previous versions of otoole, the internal read and write strategies
were exposed to the user. This was not ideal as it meant that the user
had to know the internal structure of the model. This commit adds a
public API to convert and convert_results methods. These methods
are used to convert between different file formats.

This commit also:

- adds tests for the convert and convert_results methods
- adds a new module to the otoole package called convert.py
- adds a test fixture for the super_simple model
- fixes typing errors
