+++ src/otoole/__init__.py
@@ -1,10 +1,7 @@
 # -*- coding: utf-8 -*-
 import sys
 
-from otoole.input import Context
-from otoole.read_strategies import ReadCsv, ReadDatafile, ReadExcel, ReadMemory
-from otoole.results.results import ReadCbc, ReadCplex, ReadGurobi
-from otoole.write_strategies import WriteCsv, WriteDatafile, WriteExcel
+from otoole.convert import convert, convert_results
 
 if sys.version_info[:2] >= (3, 8):
     # TODO: Import directly (no need for conditional) when `python_requires = >= 3.8`
@@ -21,17 +18,7 @@ except PackageNotFoundError:  # pragma: no cover
 finally:
     del version, PackageNotFoundError
 
+convert = convert
+convert_results = convert_results
 
-__all__ = [
-    "Context",
-    "ReadCbc",
-    "ReadCsv",
-    "ReadCplex",
-    "ReadDatafile",
-    "ReadExcel",
-    "ReadGurobi",
-    "ReadMemory",
-    "WriteCsv",
-    "WriteDatafile",
-    "WriteExcel",
-]
+__all__ = ["convert" "convert_results"]
