### 2023-06-20 ba4ac35
Added a test for cli call to convert_results
