### 2023-06-21 2e99548
Refactored WriteStrategy to remove kwargs and use class variable
