+++ src/otoole/input.py
@@ -247,13 +247,14 @@ class WriteStrategy(Strategy):
         handle = self._header()
         logger.debug(default_values)
 
+        self.input_data = inputs
         if self.write_defaults:
             try:
-                inputs = self._expand_defaults(inputs, default_values, **kwargs)
+                self.input_data = self._expand_defaults(inputs, default_values)
             except KeyError as ex:
                 logger.debug(ex)
 
-        for name, df in sorted(inputs.items()):
+        for name, df in sorted(self.input_data.items()):
             logger.debug("%s has %s columns: %s", name, len(df.index.names), df.columns)
 
             try:
@@ -278,37 +279,27 @@ class WriteStrategy(Strategy):
             handle.close()
 
     def _expand_defaults(
-        self,
-        data_to_expand: Dict[str, pd.DataFrame],
-        default_values: Dict[str, float],
-        **kwargs,
+        self, data_to_expand: Dict[str, pd.DataFrame], default_values: Dict[str, float]
     ) -> Dict[str, pd.DataFrame]:
         """Populates default value entry rows in dataframes
 
         Parameters
         ----------
-        input_data : Dict[str, pd.DataFrame],
+        data_to_expand : Dict[str, pd.DataFrame],
         default_values : Dict[str, float]
 
         Returns
         -------
-        results : Dict[str, pd.DataFrame]
-            Updated available reults dictionary
+        Dict[str, pd.DataFrame]
+            Input data with expanded default values replacing missing entries
 
         Raises
         ------
         KeyError
-            If set definitons are not in input_data and input_data is not supplied
+            If set definitions are not in input_data and input_data is not supplied
         """
 
         sets = [x for x in self.user_config if self.user_config[x]["type"] == "set"]
-
-        # if expanding results, input data is needed for set defenitions
-        if "input_data" in kwargs:
-            model_data = kwargs["input_data"]
-        else:
-            model_data = data_to_expand
-
         output = {}
         for name, data in data_to_expand.items():
             logger.info(f"Writing defaults for {name}")
@@ -320,7 +311,7 @@ class WriteStrategy(Strategy):
 
             # TODO
             # Issue with how otoole handles trade route right now.
-            # The double defenition of REGION throws an error.
+            # The double definition of REGION throws an error.
             if name == "TradeRoute":
                 output[name] = data
                 continue
@@ -329,7 +320,7 @@ class WriteStrategy(Strategy):
             index_data = {}
             for index in data.index.names:
                 try:
-                    index_data[index] = model_data[index]["VALUE"].to_list()
+                    index_data[index] = self.input_data[index]["VALUE"].to_list()
                 except KeyError as ex:
                     logger.info("Can not write default values. Supply input data")
                     raise KeyError(ex)
@@ -587,4 +578,16 @@ class ReadStrategy(Strategy):
     def read(
         self, filepath: Union[str, TextIO], **kwargs
     ) -> Tuple[Dict[str, pd.DataFrame], Dict[str, Any]]:
+        """Reads in data from file
+
+        Arguments
+        ---------
+        filepath: Union[str, TextIO]
+
+        Returns
+        -------
+        Tuple[Dict[str, pd.DataFrame], Dict[str, Any]]
+            tuple of input_data as a dictionary of pandas DataFrames and
+            dictionary of default values
+        """
         raise NotImplementedError()
