### 2023-06-21 7c6750e
Added docstrings to all strategies
