### 2023-06-21 25d6176
Factor out obtaining strategies in preparation for read and write only API functions
