+++ setup.cfg
@@ -47,7 +47,7 @@ install_requires =
     pydot
     importlib_resources; python_version<'3.7'
     pandas>=1.1
-    amply>=0.1.6
+    Amply>=0.1.6
     networkx
     flatten_dict
     openpyxl
