### 2023-06-21 c1dd1a6
Remove redundant arguments, move errors to get strategies
