+++ src/otoole/convert.py
@@ -9,7 +9,8 @@ Import the convert function from the otoole package::
 """
 import logging
 import os
-from typing import Union
+
+import pandas as pd
 
 from otoole.input import Context, ReadStrategy, WriteStrategy
 from otoole.read_strategies import ReadCsv, ReadDatafile, ReadExcel
@@ -140,10 +141,8 @@ def _get_user_config(config) -> dict:
     return user_config
 
 
-def _get_read_strategy(
-    user_config, from_format, from_path, keep_whitespace=False
-) -> Union[ReadStrategy, None]:
-    """Read OSeMOSYS parameter data from csv/datafile/excel format
+def _get_read_strategy(user_config, from_format, keep_whitespace=False) -> ReadStrategy:
+    """Get ``ReadStrategy`` for csv/datafile/excel format
 
     Arguments
     ---------
@@ -151,26 +150,23 @@ def _get_read_strategy(
         User configuration describing parameters and sets
     from_format : str
         Available options are 'datafile', 'datapackage', 'csv' and 'excel'
-    from_path : str
-        Path to destination file (if datafile or excel) or folder (csv or datapackage)
     keep_whitespace: bool, default: False
         Keep whitespace in CSVs
 
     Returns
     -------
-    dict[str, pandas.DataFrame]
-        Dictionary of pandas DataFrames containing the data
+    ReadStrategy or None
+        A ReadStrategy object. Returns None if from_format is not recognised
 
     """
     keep_whitespace = True if keep_whitespace else False
 
     if from_format == "datafile":
-        read_strategy: Union[ReadStrategy, None] = ReadDatafile(user_config=user_config)
+        read_strategy: ReadStrategy = ReadDatafile(user_config=user_config)
     elif from_format == "datapackage":
         logger.warning(
             "Reading from datapackage is deprecated, trying to read from CSVs"
         )
-        from_path = read_deprecated_datapackage(from_path)
         logger.info("Successfully read folder of CSVs")
         read_strategy = ReadCsv(
             user_config=user_config, keep_whitespace=keep_whitespace
@@ -184,22 +180,35 @@ def _get_read_strategy(
             user_config=user_config, keep_whitespace=keep_whitespace
         )  # typing: ReadStrategy
     else:
-        read_strategy = None
+        msg = f"Conversion from {from_format} is not supported"
+        raise NotImplementedError(msg)
 
     return read_strategy
 
 
-def _get_write_strategy(
-    user_config, to_format, to_path, write_defaults=False
-) -> Union[WriteStrategy, None]:
-    """ """
+def _get_write_strategy(user_config, to_format, write_defaults=False) -> WriteStrategy:
+    """Get ``WriteStrategy`` for csv/datafile/excel format
+
+    Arguments
+    ---------
+    user_config : dict
+        User configuration describing parameters and sets
+    to_format : str
+        Available options are 'datafile', 'datapackage', 'csv' and 'excel'
+    write_defaults: bool, default: False
+        Write default values to output format
+
+    Returns
+    -------
+    WriteStrategy or None
+        A ReadStrategy object. Returns None if to_format is not recognised
+
+    """
     # set write strategy
     write_defaults = True if write_defaults else False
 
     if to_format == "datapackage":
-        logger.warning("Writing to datapackage is deprecated, writing to CSVs")
-        to_path = os.path.join(os.path.dirname(to_path), "data")
-        write_strategy: Union[WriteStrategy, None] = WriteCsv(
+        write_strategy: WriteStrategy = WriteCsv(
             user_config=user_config, write_defaults=write_defaults
         )
     elif to_format == "excel":
@@ -215,7 +224,8 @@ def _get_write_strategy(
             user_config=user_config, write_defaults=write_defaults
         )
     else:
[…]
