### 2023-06-21 bb7d674
Add tests for read and write functions
