### 2023-06-21 161c619
Adds a wrapper around ReadStategy.read to return an xarray.DataSet
