+++ setup.cfg
@@ -52,6 +52,7 @@ install_requires =
     flatten_dict
     openpyxl
     pydantic
+    xarray
 [options.packages.find]
 where = src
 exclude =
