### 2023-06-26 #5511 @fneum
Excellent! This was fun reviewing.

As you can see, I already ticked all the boxes and only have a couple of questions and suggestions for you.

## Paper

- The manuscript covers all the important points.

## Repository

- The repository is in excellent condition!

## Docs

https://otoole.readthedocs.io/en/latest/functionality.html#otoole-results

- Maybe the transformation of the raw solver output file of CPLEX could be incorporated into `otoole`?

https://otoole.readthedocs.io/en/latest/data.html#parmaters-foramt

- typo "Parmaters Foramt"

https://otoole.readthedocs.io/en/latest/examples.html

- Do you need both GLPK and CBC for the example or is there a way to do it with GLPK and/or CBC alone?

https://otoole.readthedocs.io/en/latest/examples.html#view-results

- Would it be a good idea to have an `otoole clean` function to tidy up after a run?

https://otoole.readthedocs.io/en/latest/examples.html#otoole-validate

- Maybe I am missing this, but I did not generate a file `data.txt`. Hence, I could not execute this part of the examples section.

https://otoole.readthedocs.io/en/latest/examples.html#install-glpk

- You could add installation instructions for Windows (https://winglpk.sourceforge.net/).
- You could add that these solvers can be installed also via `conda`: `conda install -c conda-forge coincbc glpk` (not all combinations of operating systems and solvers).

## Functionality

- Are there any use cases where `otoole` is used within a Python script? If yes, more targeted documentation for the frontend Python API should be added besides the Module Reference.

- Is there an easy way to reduce the number of commands required to solve a model? E.g. combining building, solving and interpreting solver output?
