### 2023-07-10 00a4144
Pin pydantic to less than v2
