+++ setup.cfg
@@ -51,7 +51,7 @@ install_requires =
     networkx
     flatten_dict
     openpyxl
-    pydantic
+    pydantic<2
 [options.packages.find]
 where = src
 exclude =
