### 2023-07-10 9a7d762
Simplify dependencies for reading packaged files
