+++ src/otoole/utils.py
@@ -1,6 +1,7 @@
 import json
 import logging
 import os
+from importlib.resources import files
 from typing import Any, Dict, List, Union
 
 import pandas as pd
@@ -14,12 +15,6 @@ from otoole.preprocess.validate_config import (
     UserDefinedSet,
     UserDefinedValue,
 )
-
-try:
-    import importlib.resources as resources
-except ImportError:
-    # Try backported to PY<37 `importlib_resources`.
-    import importlib_resources as resources  # type: ignore
 
 logger = logging.getLogger(__name__)
 
@@ -42,7 +37,7 @@ def read_packaged_file(filename: str, module_name: str = None):
         with open(filename, "r") as open_file:
             contents = _read_file(open_file, ending)
     else:
-        with resources.files(module_name).joinpath(filename).open("r") as open_file:
+        with files(module_name).joinpath(filename).open("r") as open_file:
             contents = _read_file(open_file, ending)
 
     return contents
