+++ setup.cfg
@@ -45,7 +45,6 @@ install_requires =
     xlrd
     pyyaml
     pydot
-    importlib_resources; python_version<'3.7'
     pandas>=1.1
     amply>=0.1.4
     networkx
