+++ .github/workflows/python.yaml
@@ -25,7 +25,7 @@ jobs:
       - name: Lint with flake8
         run: |
           tox -e lint
-      - name: Test with pytest
+      - name: Run tests using tox
         run: |
           tox
       - name: Upload coverage data to converalls.io
