### 2023-07-10 3ef9bdc
Drop Python 3.8 from test matrix
