+++ .github/workflows/python.yaml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11"]
+        python-version: ["3.9", "3.10", "3.11"]
 
     steps:
       - uses: actions/checkout@v3
@@ -20,8 +20,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install tox
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-          pip install -e .
       - name: Lint with flake8
         run: |
           tox -e lint
