### 2023-07-10 f420d3d
Use tox config for publishing
