### 2023-07-10 b628302
Define environment variables passed to tox
