+++ .readthedocs.yml
@@ -17,7 +17,7 @@ formats:
   - pdf
 
 python:
-  version: 3.8
+  version: 3.11
   install:
     - requirements: docs/requirements.txt
     - {path: ., method: pip}
