### 2023-07-10 721e055
Update configuration for documentation
