+++ docs/requirements.txt
@@ -1,5 +1,13 @@
+amply>=0.1.4
 docutils<0.18
+flatten_dict
 Jinja2<3.1
+networkx
+openpyxl
+pandas>=1.1
+pydantic<2
+pydot
+pyyaml
 # Requirements file for ReadTheDocs, check .readthedocs.yml.
 # To build the module reference correctly, make sure every external package
 # under `install_requires` in `setup.cfg` is also listed here!
@@ -7,3 +15,5 @@ sphinx>=3.2.1
 sphinx-book-theme
 urllib3<2
 # sphinx_rtd_theme
+#otoole dependencies from setup.cfg
+xlrd
