+++ .readthedocs.yml
@@ -8,16 +8,16 @@ version: 2
 sphinx:
   configuration: docs/conf.py
 
-# Build documentation with MkDocs
-#mkdocs:
-#  configuration: mkdocs.yml
-
 # Optionally build your docs in additional formats such as PDF
 formats:
   - pdf
 
 python:
-  version: 3.11
   install:
     - requirements: docs/requirements.txt
     - {path: ., method: pip}
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.11"
