+++ setup.cfg
@@ -41,6 +41,8 @@ python_requires = >=3.9
 # Version specifiers like >=2.2,<3.0 avoid problems due to API changes in
 # new major versions. This works if the required packages follow Semantic Versioning.
 # For more information, check out https://semver.org/.
+
+# If this list changes, update docs/requirements.txt as well.
 install_requires =
     xlrd
     pyyaml
