### 2023-08-04 #5511 @trevorb1
Hi @fraukewiese, I hope you are having a good summer!

I am just checking in on the review status for this paper/repository, as it has been about 2 months since the issue was created. I think one reviewer has completed their feedback, and we are just waiting on the second one.

We were hoping to do revisions and address comments once all reviews have been submitted to streamline the process. Should we continue to wait for the second reviewer to complete their feedback? Or should we begin addressing the comments from the first fully completed review?

Thanks so much!
