### 2023-08-11 e96cfc5
glpk parsing logic and tests
