+++ tests/test_read_strategies.py
@@ -14,6 +14,7 @@ from otoole.results.results import (
     ReadCbc,
     ReadCplex,
     ReadGurobi,
+    ReadGlpk,
     check_for_duplicates,
     identify_duplicate,
     rename_duplicate_column,
@@ -610,6 +611,91 @@ class TestReadCbc:
         )
         pd.testing.assert_frame_equal(actual, expected)
 
+class TestReadGlpk:
+    """Use fixtures instead of StringIO due to the use of context managers in the logic"""
+
+    expected_sol_data = pd.DataFrame([
+            ["i", 1, "b", 3942.19479265207, 0],
+            ["i", 2, "b", 0, 0],
+            ["i", 3, "b", 0, 0],
+            ["i", 300, "b", 37.499, 0],
+            ["i", 301, "b", 31.7309999999999, 0],
+            ["j", 1, "b", 0, 0],
+            ["j", 2, "b", 0, 0],
+            ["j", 130, "l", 0, 0.282765294823514],
+            ["j", 131, "l", 0, 0.601075755990521],
+            ["j", 1025, "b", 0.0305438002923389, 0],
+            ["j", 1026, "b", 0.0422503416065477, 0],
+            ["j", 1027, "l", 0, 162679.693161095],
+            ["j", 1028, "l", 0, 81291.0524314291],
+        ], columns=["ID", "NUM", "STATUS", "PRIM", "DUAL"])
+    
+    expected_model_data = pd.DataFrame([
+            ["i", 2, "CAa4_Constraint_Capacity", "SIMPLICITY,ID,BACKSTOP1,2014"],
+            ["i", 3, "CAa4_Constraint_Capacity", "SIMPLICITY,ID,BACKSTOP1,2015"],
+            ["i", 300, "CAa4_Constraint_Capacity", "SIMPLICITY,ID,LNDFORCOV,2015"],
+            ["i", 301, "CAa4_Constraint_Capacity", "SIMPLICITY,ID,LNDFORCOV,2016"],
+            ["j", 1, "SalvageValueStorage", "SIMPLICITY,DAM,2014"],
+            ["j", 2, "SalvageValueStorage", "SIMPLICITY,DAM,2015"],
+            ["j", 130, "StorageLevelSeasonStart", "SIMPLICITY,DAM,2,2035"],
+            ["j", 131, "StorageLevelSeasonStart", "SIMPLICITY,DAM,2,2036"],
+            ["j", 1025, "NewCapacity", "SIMPLICITY,WINDPOWER,2039"],
+            ["j", 1026, "NewCapacity", "SIMPLICITY,WINDPOWER,2040"],
+            ["j", 1027, "RateOfActivity", "SIMPLICITY,ID,BACKSTOP1,1,2014"],
+            ["j", 1028, "RateOfActivity", "SIMPLICITY,IN,BACKSTOP1,1,2014"],
+        ], columns=["ID", "NUM", "NAME", "INDEX"])
+
+    def test_read_solution(self, user_config):
+        input_file = os.path.join("tests", "fixtures", "glpk_sol.txt")
+        reader = ReadGlpk(user_config)
+        actual_status, actual_data = reader.read_solution(input_file)
+        expected_status = {
+            "name":"osemosys_fast",
+            "status":"OPTIMAL",
+            "objective":4497.31967
+        }
+        assert actual_status == expected_status
+        pd.testing.assert_frame_equal(actual_data, self.expected_sol_data)
+        
+    def test_read_model(self, user_config):
+        input_file = os.path.join("tests", "fixtures", "glpk_model.txt")
+        reader = ReadGlpk(user_config)
+        actual = reader.read_model(input_file)
+        
[…]
