+++ src/otoole/results/results.py
@@ -8,6 +8,7 @@ import pandas as pd
 from otoole.input import ReadStrategy
 from otoole.preprocess.longify_data import check_datatypes
 from otoole.results.result_package import ResultsPackage
+from otoole.exceptions import OtooleError
 
 LOGGER = logging.getLogger(__name__)
 
@@ -37,7 +38,7 @@ class ReadResults(ReadStrategy):
             input_data = None
 
         available_results = self.get_results_from_file(
-            filepath, input_data
+            filepath, input_data, kwargs
         )  # type: Dict[str, pd.DataFrame]
 
         default_values = self._read_default_values(self.results_config)  # type: Dict
@@ -328,3 +329,210 @@ class ReadCbc(ReadResultsCBC):
         df["Index"] = df["Index"].str.replace(")", "", regex=False)
         df = df.drop(columns=["indexvalue"])
         return df[["Variable", "Index", "Value"]].astype({"Value": float})
+
+class ReadGlpk(ReadResultsCBC):
+    """Reads a GLPK Solution file into memory 
+    
+    The user must provide both the solution file (results.sol) and the glpk 
+    model file (model.lp). These can be generated from the following command  
+    
+    glpsol --wglp model.lp -m osemosys.txt -d simplicity.txt --write results.sol
+    
+    Arguments
+    ---------
+    user_config
+    glpk model file 
+    """
+    
+    def _convert_to_dataframe(self, glpk_model: str, glpk_sol: str) -> pd.DataFrame:
+        """Creates a wide formatted dataframe from GLPK solution
+        
+        Arguments
+        ---------
+        glpk_model: str
+            Path to GLPK model file. Can be created using the `--wglp` flag
+        glpk_sol: str
+            Path to GLPK solution file. Can be created using the `--write` flag
+            
+        Returns
+        -------
+        pd.DataFrame
+        """
+        
+        model = self.read_model(glpk_model)
+        _, sol = self.read_solution(glpk_sol)
+        return self.merge_model_sol(model, sol)
+        
+    def read_solution(self, file_path: str) -> Tuple[Dict[str,Union[str, float]], pd.DataFrame]:
+        """Reads a GLPK solution file 
+        
+        Arguments
+        ---------
+        file_path: str
+            Path to GLPK solution file. Can be created using the `--write` flag
+            
+        Returns
+        -------
+        Tuple[Dict[str,Union[str, float]], pd.DataFrame]
+            Dict[str,Union[str, float]] -> Problem name, status, and objective value 
+            pd.DataFrame -> Variables and constraints 
+            
+            
+        {"name":"osemosys", "status":"OPTIMAL", "objective":4497.31976}
+            
+           ID  NUM  STATUS PRIM DUAL
+        0  i   1    b      5    0
+        1  j   2    l      0    2
+        
+        Notes
+        -----
+        
+        -> ROWS IN SOLUTION FILE 
+        
+        i ROW ST PRIM DUAL
+        
+        ROW is the ordinal number of the row
+        ST is one of:
+            b = inactive constraint;
+            l = inequality constraint active on its lower bound;
+            u = inequality constraint active on its upper bound;
+            f = active free (unounded) row;
+            s = active equality constraint.
+        PRIM specifies the row primal value (float)
+        DUAL specifies the row dual value (float)
+        
+        -> COLUMNS IN SOLUTION FILE
+        
+        j COL ST PRIM DUAL
+
+        COL specifies the column ordinal number
+        ST contains one of the following lower-case letters that specifies the column status in the basic solution:
+            b = basic variable
+            l = non-basic variable having its lower bound active
+            u = non-basic variable having its upper bound active
[…]
