+++ src/otoole/results/results.py
@@ -5,10 +5,10 @@ from typing import Any, Dict, List, Set, TextIO, Tuple, Union
 
 import pandas as pd
 
+from otoole.exceptions import OtooleError
 from otoole.input import ReadStrategy
 from otoole.preprocess.longify_data import check_datatypes
 from otoole.results.result_package import ResultsPackage
-from otoole.exceptions import OtooleError
 
 LOGGER = logging.getLogger(__name__)
 
@@ -38,7 +38,8 @@ class ReadResults(ReadStrategy):
             input_data = None
 
         available_results = self.get_results_from_file(
-            filepath, input_data)  # type: Dict[str, pd.DataFrame]
+            filepath, input_data
+        )  # type: Dict[str, pd.DataFrame]
 
         default_values = self._read_default_values(self.results_config)  # type: Dict
 
@@ -329,70 +330,80 @@ class ReadCbc(ReadResultsCBC):
         df = df.drop(columns=["indexvalue"])
         return df[["Variable", "Index", "Value"]].astype({"Value": float})
 
+
 class ReadGlpk(ReadResultsCBC):
-    """Reads a GLPK Solution file into memory 
-    
-    The user must provide both the solution file (results.sol) and the glpk 
-    model file (model.lp) to generate the complete solution.   
-    
+    """Reads a GLPK Solution file into memory
+
+    The user must provide both the solution file (results.sol) and the glpk
+    model file (model.lp) to generate the complete solution.
+
     glpsol --wglp model.lp -m osemosys.txt -d simplicity.txt --write results.sol
     """
-    
+
     def __init__(self, user_config: Dict[str, Dict], glpk_model: str = None):
         """
         glpk_model: str
             Path to GLPK model file. Can be created using the `--wglp` flag.
-            If not provided, the solution file will be processed without 
-            corresponding english names or index defenitions.  
+            If not provided, the solution file will be processed without
+            corresponding english names or index defenitions.
         """
         super().__init__(user_config)
         self.glpk_model = glpk_model
-    
-    def _convert_to_dataframe(self, glpk_sol: str) -> pd.DataFrame:
+
+    def _convert_to_dataframe(self, glpk_sol: Union[str, TextIO]) -> pd.DataFrame:
         """Creates a wide formatted dataframe from GLPK solution
-        
+
         Arguments
         ---------
         glpk_sol: str
             Path to GLPK solution file. Can be created using the `--write` flag
-            
+
         Returns
         -------
         pd.DataFrame
         """
-        
+
         model = self.read_model()
-        _, sol = self.read_solution(glpk_sol)
+
+        if isinstance(glpk_sol, str):
+            with open(glpk_sol, "r") as sol_file:
+                _, sol = self.read_solution(sol_file)
+        elif isinstance(glpk_sol, StringIO):
+            _, sol = self.read_solution(glpk_sol)
+        else:
+            raise TypeError("Argument filepath type must be a string or an open file")
+
         return self.merge_model_sol(model, sol)
-        
-    def read_solution(self, file_path: str) -> Tuple[Dict[str,Union[str, float]], pd.DataFrame]:
-        """Reads a GLPK solution file 
-        
+
+    def read_solution(
+        self, file_path: Union[str, TextIO]
+    ) -> Tuple[Dict[str, Union[str, float]], pd.DataFrame]:
+        """Reads a GLPK solution file
+
         Arguments
         ---------
         file_path: str
             Path to GLPK solution file. Can be created using the `--write` flag
-            
+
         Returns
         -------
         Tuple[Dict[str,Union[str, float]], pd.DataFrame]
-            Dict[str,Union[str, float]] -> Problem name, status, and objective value 
-            pd.DataFrame -> Variables and constraints 
-            
-            
[…]
