### 2023-08-12 464241b
update to pydantic 2
