+++ docs/requirements.txt
@@ -5,7 +5,7 @@ Jinja2<3.1
 networkx
 openpyxl
 pandas>=1.1
-pydantic<2
+pydantic>=2
 pydot
 pyyaml
 # Requirements file for ReadTheDocs, check .readthedocs.yml.
