+++ setup.cfg
@@ -52,7 +52,7 @@ install_requires =
     networkx
     flatten_dict
     openpyxl
-    pydantic<2
+    pydantic>=2
 [options.packages.find]
 where = src
 exclude =
