+++ docs/data.rst
@@ -68,7 +68,7 @@ Sets are defined as follows::
 .. NOTE::
    It's convention in OSeMOSYS to capitalize set names
 
-Parmaters Foramt
+Parameters Format
 ~~~~~~~~~~~~~~~~
 
 Parameters are defined as follows. When referencing set indices use the full
@@ -112,7 +112,7 @@ repository for a complete example.
         dtype: str
         type: set
 
-2. Parmater definition of ``AccumulatedAnnualDemand``::
+2. Parameter definition of ``AccumulatedAnnualDemand``::
 
     AccumulatedAnnualDemand:
         short_name: AccAnnualDemand
