### 2023-08-14 b844570
update install instructions
