+++ docs/examples.rst
@@ -15,10 +15,86 @@ functionality in seperate simple use cases.
         git clone https://github.com/OSeMOSYS/simplicity.git
         cd simplicity
 
-.. CAUTION::
-    While ``otoole`` does not require a solver, these examples
-    will use the free and open source solvers GLPK_ and CBC_.
-    Installation instructions are described in the `Solver Setup`_ section.
+Solver Setup
+------------
+
+Objective
+~~~~~~~~~
+
+Install GLPK_ (required) and CBC_ (optional) to use in the otoole examples.
+While ``otoole`` does not require a solver, these examples will use the free
+and open source solvers GLPK_ and CBC_.
+
+1. Install GLPK
+~~~~~~~~~~~~~~~~
+
+GLPK_ is a free and open-source linear program solver. Full
+install instructions can be found on the `GLPK Website`_, however, the
+abbreviated instructions are shown below
+
+To install GLPK on **Linux**, run the command::
+
+    sudo apt-get update
+    sudo apt-get install glpk glpk-utils
+
+To install GLPK on **Mac**, run the command::
+
+    brew install glpk
+
+To install GLPK on **Windows**, follow the instructions on the
+`GLPK Website`_. Be sure to add GLPK to
+your environment variables after installation
+
+Alternatively, if you use Anaconda_ to manage
+your Python packages, you can install GLPK via the command::
+
+    conda install -c conda-forge glpk
+
+2. Test the GLPK install
+~~~~~~~~~~~~~~~~~~~~~~~~
+Once installed, you should be able to call the ``glpsol`` command::
+
+    $ glpsol
+    GLPSOL: GLPK LP/MIP Solver, v4.65
+    No input problem file specified; try glpsol --help
+
+3. Install CBC
+~~~~~~~~~~~~~~
+
+CBC_ is a free and open-source mixed integer linear programming solver. Full
+install instructions can be found on the CBC_ website, however, the abbreviated
+instructions are shown below
+
+To install CBC on **Linux**, run the command::
+
+    sudo apt-get install coinor-cbc coinor-libcbc-dev
+
+To install CBC on **Mac**, run the command::
+
+    brew install coin-or-tools/coinor/cbc
+
+To install CBC on **Windows**, follow the install instruction on the CBC_
+website.
+
+Alternatively, if you use Anaconda_ to manage
+your Python packages, you can install CBC via the command::
+
+    conda install -c conda-forge coincbc
+
+4. Test the CBC install
+~~~~~~~~~~~~~~~~~~~~~~~
+Once installed, you should be able to directly call CBC::
+
+    $ cbc
+    Welcome to the CBC MILP Solver
+    Version: 2.10.3
+    Build Date: Mar 24 2020
+
+    CoinSolver takes input from arguments ( - switches to stdin)
+    Enter ? for list of commands or help
+    Coin:
+
+You can exit the solver by typing ``quit``
 
 Data Conversion with CSVs
 -------------------------
@@ -211,13 +287,18 @@ codes are shown in bold face.
 
 .. image:: _static/validataion_model.png
 
-1. Create the Validation File
+1. Download the example datafile
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+The MathProg datafile describing this model can be found on the
+:ref:`examples-validation` page. Download the file and save it as ``data.txt``
+
+2. Create the Validation File
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Create a configuration validation ``yaml`` file::
 
     $ touch validate.yaml
 
-2. Create ``FUEL`` Codes
+3. Create ``FUEL`` Codes
 ~~~~~~~~~~~~~~~~~~~~~~~~
[…]
