+++ docs/examples-validation.rst
@@ -1,0 +1,213 @@
+:orphan:
+
+.. _examples-validation:
+
+-----------------------
+Example Validation File 
+-----------------------
+
+This page holds the datafile used in the validation example. The file can 
+either be copy/pasted from below, or directly downloaded from :download:`here <_static/validation-data.txt>` ::
+
+    # Model file written by *otoole*
+    param default 0 : AccumulatedAnnualDemand :=
+    ;
+    param default -1 : AnnualEmissionLimit :=
+    ;
+    param default 0 : AnnualExogenousEmission :=
+    ;
+    param default 1 : AvailabilityFactor :=
+    ;
+    param default 1 : CapacityFactor :=
+    ;
+    param default 0 : CapacityOfOneTechnologyUnit :=
+    ;
+    param default 1 : CapacityToActivityUnit :=
+    R PWRWND 31.536
+    R PWRCOA 31.536
+    R TRNELC 31.536
+    ;
+    param default 0 : CapitalCost :=
+    R PWRWND 2020 1500
+    R PWRWND 2021 1500
+    R PWRWND 2022 1500
+    R PWRCOA 2020 5000
+    R PWRCOA 2021 5000
+    R PWRCOA 2022 5000
+    ;
+    param default 0 : CapitalCostStorage :=
+    ;
+    param default 0 : Conversionld :=
+    ;
+    param default 0 : Conversionlh :=
+    ;
+    param default 0 : Conversionls :=
+    ;
+    set DAILYTIMEBRACKET :=
+    ;
+    set DAYTYPE :=
+    ;
+    param default 0.00137 : DaySplit :=
+    ;
+    param default 7 : DaysInDayType :=
+    ;
+    param default 1 : DepreciationMethod :=
+    ;
+    param default 0.05 : DiscountRate :=
+    ;
+    param default 0.05 : DiscountRateStorage :=
+    ;
+    set EMISSION :=
+    ;
+    param default 0 : EmissionActivityRatio :=
+    ;
+    param default 0 : EmissionsPenalty :=
+    ;
+    set FUEL :=
+    WND00
+    COA00
+    ELC01
+    ELC02
+    ;
+    param default 0 : FixedCost :=
+    ;
+    param default 0 : InputActivityRatio :=
+    R PWRWND WND00 1 2020 1
+    R PWRWND WND00 1 2021 1
+    R PWRWND WND00 1 2022 1
+    R PWRCOA COA00 1 2020 1
+    R PWRCOA COA00 1 2021 1
+    R PWRCOA COA00 1 2022 1
+    R TRNELC ELC01 1 2020 1
+    R TRNELC ELC01 1 2021 1
+    R TRNELC ELC01 1 2022 1
+    ;
+    set MODE_OF_OPERATION :=
+    1
+    ;
+    param default 0 : MinStorageCharge :=
+    ;
+    param default -1 : ModelPeriodEmissionLimit :=
+    ;
+    param default 0 : ModelPeriodExogenousEmission :=
+    ;
+    param default 1 : OperationalLife :=
+    R PWRWND 20
+    R PWRCOA 30
+    ;
+    param default 0 : OperationalLifeStorage :=
+    ;
+    param default 0 : OutputActivityRatio :=
+    R MINWND WND00 1 2020 1
+    R MINWND WND00 1 2021 1
+    R MINWND WND00 1 2022 1
+    R MINCOA COA00 1 2020 1
+    R MINCOA COA00 1 2021 1
+    R MINCOA COA00 1 2022 1
+    R PWRWND ELC01 1 2020 1
[…]
