### 2023-08-15 3758c44
Refactor cli functions to use python interface
