### 2023-08-15 c1f8865
update ReadGlpk logic
