+++ src/otoole/results/results.py
@@ -340,15 +340,22 @@ class ReadGlpk(ReadResultsCBC):
     glpsol --wglp model.lp -m osemosys.txt -d simplicity.txt --write results.sol
     """
 
-    def __init__(self, user_config: Dict[str, Dict], glpk_model: str = None):
+    def __init__(self, user_config: Dict[str, Dict], glpk_model: Union[str, TextIO]):
         """
-        glpk_model: str
+        glpk_model: Union[str, TextIO]
             Path to GLPK model file. Can be created using the `--wglp` flag.
             If not provided, the solution file will be processed without
             corresponding english names or index defenitions.
         """
         super().__init__(user_config)
-        self.glpk_model = glpk_model
+
+        if isinstance(glpk_model, str):
+            with open(glpk_model, "r") as model_file:
+                self.model = self.read_model(model_file)
+        elif isinstance(glpk_model, StringIO):
+            self.model = self.read_model(glpk_model)
+        else:
+            raise TypeError("Argument filepath type must be a string or an open file")
 
     def _convert_to_dataframe(self, glpk_sol: Union[str, TextIO]) -> pd.DataFrame:
         """Creates a wide formatted dataframe from GLPK solution
@@ -363,17 +370,62 @@ class ReadGlpk(ReadResultsCBC):
         pd.DataFrame
         """
 
-        model = self.read_model()
-
         if isinstance(glpk_sol, str):
-            with open(glpk_sol, "r") as sol_file:
-                _, sol = self.read_solution(sol_file)
+            with open(glpk_sol, "r"):
+                _, sol = self.read_solution(glpk_sol)
         elif isinstance(glpk_sol, StringIO):
             _, sol = self.read_solution(glpk_sol)
         else:
             raise TypeError("Argument filepath type must be a string or an open file")
 
-        return self.merge_model_sol(model, sol)
+        return self._merge_model_sol(sol)
+
+    def read_model(self, file_path: Union[str, TextIO]) -> pd.DataFrame:
+        """Reads in a GLPK Model File
+
+        Arguments
+        ---------
+        file_path: str
+            Path to GLPK model file. Can be created using the `--wglp` flag.
+
+        Returns
+        -------
+        pd.DataFrame
+
+           ID  NUM  NAME                      INDEX
+        0  i   1    CAa4_Constraint_Capacity  "SIMPLICITY,ID,BACKSTOP1,2015"
+        1  j   2    NewCapacity               "SIMPLICITY,WINDPOWER,2039"
+
+        Notes
+        -----
+
+        -> GENERAL LAYOUT OF SOLUTION FILE
+
+        n p NAME # p = problem instance
+        n z NAME # z = objective function
+        n i ROW NAME # i = constraint name, ROW is the row ordinal number
+        n j COL NAME # j = variable name, COL is the column ordinal number
+        """
+
+        df = pd.read_csv(
+            file_path,
+            header=None,
+            sep=r"\s+",
+            index_col=0,
+            names=["ID", "NUM", "value", 4, 5],
+        ).drop(columns=[4, 5])
+
+        df = df[(df["ID"].isin(["i", "j"])) & (df["value"] != "cost")]
+
+        df[["NAME", "INDEX"]] = df["value"].str.split("[", expand=True)
+        df["INDEX"] = df["INDEX"].map(lambda x: x.split("]")[0])
+        df = (
+            df[["ID", "NUM", "NAME", "INDEX"]]
+            .astype({"ID": str, "NUM": int, "NAME": str, "INDEX": str})
+            .reset_index(drop=True)
+        )
+
+        return df
 
     def read_solution(
         self, file_path: Union[str, TextIO]
@@ -429,99 +481,36 @@ class ReadGlpk(ReadResultsCBC):
[…]
