+++ tests/test_read_strategies.py
@@ -615,6 +615,31 @@ class TestReadCbc:
 class TestReadGlpk:
     """Use fixtures instead of StringIO due to the use of context managers in the logic"""
 
+    model_data = dedent(
+        """p lp min 12665 9450 82606
+n p osemosys_fast
+n z cost
+i 1 f
+n i 1 cost
+i 2 u -0
+n i 2 CAa4_Constraint_Capacity[SIMPLICITY,ID,BACKSTOP1,2014]
+i 3 u -0
+n i 3 CAa4_Constraint_Capacity[SIMPLICITY,ID,BACKSTOP1,2015]
+i 300 u 147.115
+n i 300 CAa4_Constraint_Capacity[SIMPLICITY,ID,LNDFORCOV,2015]
+i 301 u 144.231
+n i 301 CAa4_Constraint_Capacity[SIMPLICITY,ID,LNDFORCOV,2016]
+n j 1 SalvageValueStorage[SIMPLICITY,DAM,2014]
+n j 2 SalvageValueStorage[SIMPLICITY,DAM,2015]
+n j 130 StorageLevelSeasonStart[SIMPLICITY,DAM,2,2035]
+n j 131 StorageLevelSeasonStart[SIMPLICITY,DAM,2,2036]
+n j 1025 NewCapacity[SIMPLICITY,WINDPOWER,2039]
+n j 1026 NewCapacity[SIMPLICITY,WINDPOWER,2040]
+n j 1027 RateOfActivity[SIMPLICITY,ID,BACKSTOP1,1,2014]
+n j 1028 RateOfActivity[SIMPLICITY,IN,BACKSTOP1,1,2014]
+"""
+    )
+
     sol_data = dedent(
         """c Problem:    osemosys_fast
 c Rows:       12665
@@ -636,71 +661,97 @@ j 131 l 0 0.601075755990521
 j 1025 b 0.0305438002923389 0
 j 1026 b 0.0422503416065477 0
 j 1027 l 0 162679.693161095
-j 1028 l 0 81291.0524314291"""
+j 1028 l 0 81291.0524314291
+e o f
+"""
     )
 
-    expected_sol_data = pd.DataFrame(
-        [
-            ["i", 1, "b", 3942.19479265207, 0],
-            ["i", 2, "b", 0, 0],
-            ["i", 3, "b", 0, 0],
-            ["i", 300, "b", 37.499, 0],
-            ["i", 301, "b", 31.7309999999999, 0],
-            ["j", 1, "b", 0, 0],
-            ["j", 2, "b", 0, 0],
-            ["j", 130, "l", 0, 0.282765294823514],
-            ["j", 131, "l", 0, 0.601075755990521],
-            ["j", 1025, "b", 0.0305438002923389, 0],
-            ["j", 1026, "b", 0.0422503416065477, 0],
-            ["j", 1027, "l", 0, 162679.693161095],
-            ["j", 1028, "l", 0, 81291.0524314291],
-        ],
-        columns=["ID", "NUM", "STATUS", "PRIM", "DUAL"],
-    )
+    def test_read_model(self, user_config):
+        model_data = self.model_data
+        with StringIO(model_data) as file_buffer:
+            reader = ReadGlpk(user_config=user_config, glpk_model=file_buffer)
+        actual = reader.model
 
-    expected_model_data = pd.DataFrame(
-        [
-            ["i", 2, "CAa4_Constraint_Capacity", "SIMPLICITY,ID,BACKSTOP1,2014"],
-            ["i", 3, "CAa4_Constraint_Capacity", "SIMPLICITY,ID,BACKSTOP1,2015"],
[…]
