### 2023-08-17 fba3145
Add a read_results function to Python interface
