+++ src/otoole/convert.py
@@ -9,30 +9,72 @@ Import the convert function from the otoole package::
 """
 import logging
 import os
-from typing import Dict, Tuple
+from typing import Dict, Optional, Tuple, Union
 
 import pandas as pd
 
 from otoole.input import Context, ReadStrategy, WriteStrategy
 from otoole.read_strategies import ReadCsv, ReadDatafile, ReadExcel
-from otoole.results.results import ReadCbc, ReadCplex, ReadGurobi
+from otoole.results.results import ReadCbc, ReadCplex, ReadGurobi, ReadResults
 from otoole.utils import _read_file, read_deprecated_datapackage, validate_config
 from otoole.write_strategies import WriteCsv, WriteDatafile, WriteExcel
 
 logger = logging.getLogger(__name__)
 
 
+def read_results(
+    config: str, from_format: str, from_path: str, input_format: str, input_path: str
+) -> Tuple[Dict[str, pd.DataFrame], Dict[str, float]]:
+    """Read OSeMOSYS results from CBC, GLPK or Gurobi results files
+
+    Arguments
+    ---------
+    config : str
+        Path to config file
+    from_format : str
+        Available options are 'datafile', 'csv', 'excel' and 'datapackage' [deprecated]
+    from_path : str
+        Path to source file (if datafile or excel) or folder (csv)
+    input_format: str
+        Format of input data. Available options are 'datafile', 'csv' and 'excel'
+    input_path: str
+        Path to input data    input_format: str
+        Format of input data. Available options are 'datafile', 'csv' and 'excel'
+    input_path: str
+        Path to input data
+
+    Returns
+    -------
+    Tuple[dict[str, pd.DataFrame], dict[str, float]]
+        Dictionary of parameter and set data and dictionary of default values
+    """
+    user_config = _get_user_config(config)
+    input_strategy = _get_read_strategy(user_config, input_format)
+    result_strategy = _get_read_result_strategy(user_config, from_format)
+
+    if input_strategy:
+        input_data, _ = input_strategy.read(input_path)
+    else:
+        input_data = {}
+
+    if result_strategy:
+        results, default_values = result_strategy.read(from_path, input_data=input_data)
+        return results, default_values
+    else:
+        msg = "Conversion from {} is not yet implemented".format(from_format)
+        raise NotImplementedError(msg)
+
+
 def convert_results(
-    config,
-    from_format,
-    to_format,
-    from_path,
-    to_path,
-    input_datapackage=None,
-    input_csvs=None,
-    input_datafile=None,
+    config: str,
+    from_format: str,
+    to_format: str,
+    from_path: str,
+    to_path: str,
+    input_format: str,
+    input_path: str,
     write_defaults=False,
-):
+) -> bool:
     """Post-process results from a CBC, CPLEX or Gurobi solution file into CSV format
 
     Arguments
@@ -47,12 +89,10 @@ def convert_results(
         Path to cbc, cplex or gurobi solution file
     to_path : str
         Path to destination folder
-    input_datapackage : str
-        Path to folder containing datapackage.json
-    input_csvs : str
-        Path to folder containing CSVs
-    input_datafile : str
-        Path to datafile
+    input_format: str
+        Format of input data. Available options are 'datafile', 'csv' and 'excel'
+    input_path: str
+        Path to input data
     write_defaults : str
         Write default values to CSVs
 
@@ -66,25 +106,11 @@ def convert_results(
         from_format, to_format
     )
 
-    read_strategy = None
-    write_strategy = None
-
-    if config:
-        _, ending = os.path.splitext(config)
-        with open(config, "r") as config_file:
-            user_config = _read_file(config_file, ending)
[…]
