### 2023-08-18 6b058d0
Resolves final issues raised in PR
