### 2023-08-18 9a42fb6
minor doc fixes
