+++ src/otoole/results/results.py
@@ -6,7 +6,8 @@ from typing import Any, Dict, List, Set, TextIO, Tuple, Union
 import pandas as pd
 
 from otoole.input import ReadStrategy
-from otoole.preprocess.longify_data import check_datatypes
+
+# from otoole.preprocess.longify_data import check_datatypes
 from otoole.results.result_package import ResultsPackage
 
 LOGGER = logging.getLogger(__name__)
@@ -179,91 +180,109 @@ def rename_duplicate_column(index: List) -> List:
     return column
 
 
-class ReadCplex(ReadResults):
-    """ """
+# class ReadCplex(ReadResults):
+#     """ """
 
-    def get_results_from_file(
-        self, filepath: Union[str, TextIO], input_data
-    ) -> Dict[str, pd.DataFrame]:
+#     def get_results_from_file(
+#         self, filepath: Union[str, TextIO], input_data
+#     ) -> Dict[str, pd.DataFrame]:
 
-        if input_data:
-            years = input_data["YEAR"].values  # type: List
-            start_year = int(years[0])
-            end_year = int(years[-1])
-        else:
-            raise RuntimeError("To process CPLEX results please provide the input file")
+#         if input_data:
+#             years = input_data["YEAR"].values  # type: List
+#             start_year = int(years[0])
+#             end_year = int(years[-1])
+#         else:
+#             raise RuntimeError("To process CPLEX results please provide the input file")
 
-        if isinstance(filepath, str):
-            with open(filepath, "r") as sol_file:
-                data = self.extract_rows(sol_file, start_year, end_year)
-        elif isinstance(filepath, StringIO):
-            data = self.extract_rows(filepath, start_year, end_year)
-        else:
-            raise TypeError("Argument filepath type must be a string or an open file")
+#         if isinstance(filepath, str):
+#             with open(filepath, "r") as sol_file:
+#                 data = self.extract_rows(sol_file, start_year, end_year)
+#         elif isinstance(filepath, StringIO):
+#             data = self.extract_rows(filepath, start_year, end_year)
+#         else:
+#             raise TypeError("Argument filepath type must be a string or an open file")
 
-        results = {}
+#         results = {}
 
-        for name in data.keys():
-            results[name] = self.convert_df(data[name], name, start_year, end_year)
+#         for name in data.keys():
+#             results[name] = self.convert_df(data[name], name, start_year, end_year)
 
-        return results
+#         return results
 
-    def extract_rows(
-        self, sol_file: TextIO, start_year: int, end_year: int
-    ) -> Dict[str, List[List[str]]]:
-        """ """
-        data = {}  # type: Dict[str, List[List[str]]]
-        for linenum, line in enumerate(sol_file):
-            line = line.replace("\n", "")
-            try:
-                row_as_list = line.split("\t")  # type: List[str]
-                name = row_as_list[0]  # type: str
+#     def extract_rows(
+#         self, sol_file: TextIO, start_year: int, end_year: int
+#     ) -> Dict[str, List[List[str]]]:
+#         """ """
+#         data = {}  # type: Dict[str, List[List[str]]]
+#         for linenum, line in enumerate(sol_file):
+#             line = line.replace("\n", "")
+#             try:
+#                 row_as_list = line.split("\t")  # type: List[str]
+#                 name = row_as_list[0]  # type: str
 
-                if name in data.keys():
-                    data[name].append(row_as_list)
-                else:
-                    data[name] = [row_as_list]
-            except ValueError as ex:
-                msg = "Error caused at line {}: {}. {}"
-                raise ValueError(msg.format(linenum, line, ex))
-        return data
[…]
