### 2023-08-22 c413ae6
updated ReadCplex tests
