### 2023-08-22 53871a8
update cli example docs
