+++ docs/examples.rst
@@ -6,7 +6,7 @@ Examples
 
 This page will present examples to show the full functionality of ``otoole``. It will
 walk through the ``convert``, ``results``, ``setup``, ``viz`` and ``validate``
-functionality in seperate simple use cases.
+functionality in separate simple use cases.
 
 .. NOTE::
     To follow these examples, clone the Simplicity_ repository and run all commands
@@ -34,12 +34,12 @@ abbreviated instructions are shown below
 
 To install GLPK on **Linux**, run the command::
 
-    sudo apt-get update
-    sudo apt-get install glpk glpk-utils
+    $ sudo apt-get update
+    $ sudo apt-get install glpk glpk-utils
 
 To install GLPK on **Mac**, run the command::
 
-    brew install glpk
+    $ brew install glpk
 
 To install GLPK on **Windows**, follow the instructions on the
 `GLPK Website`_. Be sure to add GLPK to
@@ -48,7 +48,7 @@ your environment variables after installation
 Alternatively, if you use Anaconda_ to manage
 your Python packages, you can install GLPK via the command::
 
-    conda install -c conda-forge glpk
+    $ conda install -c conda-forge glpk
 
 2. Test the GLPK install
 ~~~~~~~~~~~~~~~~~~~~~~~~
@@ -57,6 +57,9 @@ Once installed, you should be able to call the ``glpsol`` command::
     $ glpsol
     GLPSOL: GLPK LP/MIP Solver, v4.65
     No input problem file specified; try glpsol --help
+
+.. TIP::
+    See the `GLPK Wiki`_ for more information on the ``glpsol`` command
 
 3. Install CBC
 ~~~~~~~~~~~~~~
@@ -67,11 +70,11 @@ instructions are shown below
 
 To install CBC on **Linux**, run the command::
 
-    sudo apt-get install coinor-cbc coinor-libcbc-dev
+    $ sudo apt-get install coinor-cbc coinor-libcbc-dev
 
 To install CBC on **Mac**, run the command::
 
-    brew install coin-or-tools/coinor/cbc
+    $ brew install coin-or-tools/coinor/cbc
 
 To install CBC on **Windows**, follow the install instruction on the CBC_
 website.
@@ -79,7 +82,7 @@ website.
 Alternatively, if you use Anaconda_ to manage
 your Python packages, you can install CBC via the command::
 
-    conda install -c conda-forge coincbc
+    $ conda install -c conda-forge coincbc
 
 4. Test the CBC install
 ~~~~~~~~~~~~~~~~~~~~~~~
@@ -96,122 +99,103 @@ Once installed, you should be able to directly call CBC::
 
 You can exit the solver by typing ``quit``
 
-Data Conversion with CSVs
--------------------------
+Input Data Conversion
+---------------------
 
 Objective
 ~~~~~~~~~
 
-Use a folder of CSV data to build and solve an OSeMOSYS model with CBC_. Generate
-the full suite of OSeMOSYS results.
+Convert input data between CSV, Excel, and GNU MathProg data formats.
 
-1. ``otoole`` Convert
-~~~~~~~~~~~~~~~~~~~~~
-We first want to convert the folder of Simplicity_ CSVs into
-an OSeMOSYS datafile called ``simplicity.txt``::
+1. Clone ``Simplicity``
+~~~~~~~~~~~~~~~~~~~~~~~
+If not already done so, clone the Simplicity_ repository::
+
+    $ git clone https://github.com/OSeMOSYS/simplicity.git
+    $ cd simplicity
+
+.. NOTE::
+    Further information on the ``config.yaml`` file is in the :ref:`template-setup` section
+
+2. Convert CSV data into MathProg data
[…]
