### 2023-08-29 OSeMOSYS/otoole#192 @trevorb1
[FEATURE]: Create a central configuration file 
### Feature Request

We are beginning to end up with a lot of configuration options for all the different "features" of `otoole`, including validation, parameter and set definition, results, writing defaults, etc.  It may be beneficial to harmonise all of these options into a single configuration file.

In addition to this simplifying the amount of arguments the user has to pass into `otoole` functions, this can provide an avenue for more extensive features to be added. For example:
- Create standard (configurable) plots of input and result data
- Automate the process of `convert data -> build model -> solve model -> process results`, through an `otoole run` command. In this process, the user could select a solver (with whatever solver options they like) and apply any OSeMOSYS GitHub hosted file they like (ie. the commonly used [preprocessing script](https://github.com/OSeMOSYS/OSeMOSYS_GNU_MathProg/blob/master/scripts/preprocess_data.py)) via this central configuration file.
- Remove intermediate files the user does not need to interact with (`*.lp`, `*.glp`, `*.sol`, ect...), either automatically in the `otoole run` command, or via a separate `otoole clean`. The user can specify what file types should and should not be removed.
- ...

### Is your feature related to a bug?

_No response_

### Suggested Solution

_No response_

### Additional Info

_No response_
