### 2023-08-31 #5511 @trevorb1
Hi @fneum; thank you again for the detailed review of `otoole`! We really appreciate it. Please find below point-by-point responses to all of your comments. If any response is unclear, needs further explanation/discussion, or requires further attention, please let us know. Thank you!

**The manuscript covers all the important points.**
Thank you! The manuscript has been updated to reflect all changes addressed in this review (see [this commit](https://github.com/OSeMOSYS/otoole/commit/ce79061c803f9cd660c52af5a6258de4dd72e7fb))

**The repository is in excellent condition!**
Thank you :)

**Maybe the transformation of the raw solver output file of CPLEX could be incorporated into otoole?**
Agreed. `otoole` can now process raw output CPLEX solution files. The change has been implemented in [this PR](https://github.com/OSeMOSYS/otoole/pull/190)  and documented [here](https://otoole.readthedocs.io/en/latest/examples.html#process-a-solution-from-cplex)

**typo "Parmaters Foramt"**
Thank you for noticing this! This has been corrected in [this PR](https://github.com/OSeMOSYS/otoole/pull/189)

**Do you need both GLPK and CBC for the example or is there a way to do it with GLPK and/or CBC alone?**
Good point on being able to use `otoole` with only one solver, as this may reduce the barrier of entry for new users. Since many OSeMOSYS workflows use GLPK to build the model file, we have added functionality to process GLPK solution files. We do recognize this may be of limited use, due to the poor performance of GLPK with large models, but agree that it can reduce the learning curve for `otoole`.

The change has been implemented in [this PR](https://github.com/OSeMOSYS/otoole/pull/187) and documented [here](https://otoole.readthedocs.io/en/latest/examples.html#process-a-solution-from-glpk)

 **Would it be a good idea to have an `otoole clean` function to tidy up after a run?**
This is a great idea, which we think relates closely to your `otoole run` idea. We have created a [new issue ticket](https://github.com/OSeMOSYS/otoole/issues/192) for this, however, have not implemented it.

The current scope of `otoole` is just the pre and post processing of data. Bridging this gap will require serious development thought on how to implement functions that remain useful to the OSeMOSYS community. Therefore, within the current scope of the project, we leave the removal of intermediate files up to the user. Please note, as elaborated in the issue ticket, in the long term it would be great to have a central configuration file for `otoole` to handle a wide array of options – including the automatic removal of intermediate files.

**Maybe I am missing this, but I did not generate a file `data.txt`. Hence, I could not execute this part of the examples section.**
Thank you for flagging this! The data file was mistakenly not included with the documentation. This has been corrected in [this PR](https://github.com/OSeMOSYS/otoole/pull/189), and the validation instructions have been updated [here](https://otoole.readthedocs.io/en/latest/examples.html#model-validation).

**You could add [solver] installation instructions for Windows (https://winglpk.sourceforge.net/). You could add that these solvers can be installed also via conda: `conda install -c conda-forge coincbc glpk` (not all combinations of operating systems and solvers).**
Agreed. Solver download instructions have been updated to reflect Windows operating systems and users who use `conda`. The change has been implemented in [this PR](https://github.com/OSeMOSYS/otoole/pull/189) and the instructions have been updated [here](https://otoole.readthedocs.io/en/latest/examples.html#solver-setup)

**Are there any use cases where `otoole` is used within a Python script? If yes, more targeted documentation for the front end Python API should be added besides the Module Reference.**
Yes, some users use `otoole` directly in Python scripts, rather than through the command line. We agree that documenting this process better is beneficial.  A public python API has been added to `otoole` to allow users to convert data, read and write data, and read results without having to know the internal data structures of otoole. Moreover, dedicated examples on how to use the Python API have been added to the documentation website.

The update has been implemented in [this PR](https://github.com/OSeMOSYS/otoole/pull/180) and documented [here](https://otoole.readthedocs.io/en/latest/convert.html)

**Is there an easy way to reduce the number of commands required to solve a model? E.g. combining building, solving and interpreting solver output?**
This is a good thought! Similar to the `otoole clean` discussion, we view the current scope of `otoole` to be only the pre and post processing data. As further discussed in [this issue](https://github.com/OSeMOSYS/otoole/issues/192), creating a central configuration file where to users can launch model runs, specify solvers, and apply external scripts, is a long-term goal that requires serious development thought outside the current scope of `otoole`.
