### 2023-08-31 #5511 @trevorb1
Hi @fraukewiese. A couple quick questions on the review process:
1. With the changes addressing @fneum's comments, we have now released `otoole` version `1.1.0`. We originally submitted `otoole` under version `1.0.3`. Do we have to indicate this bump in version in anyway?
2. In the paper, we left the date as the original submission date (May 3, 2023). Does this need to be updated as we revise the paper in the review?

Thank you!
