### 2023-09-11 #5511 @olejandro
@fraukewiese apologies for the delay.
@trevorb1 below are comments on the software paper. I'll come back shortly on the rest.

The paper is concise and generally well-written. I believe it covers most of what is required by JOSS. What is missing is a description of the state of the field. Currently it focuses solely on the OSeMOSYS users, however the problems that it is attempting to solve are common to other energy modelling communities. How does `otoole`, which is developed for `OseMOSYS`, differ from similar tools that were developed for other energy system models / model generators / frameworks?

The following sections require minor clarification / edits.

**Statement of need**
- `is a highly cited and reputable open-source framework` - please provide a reference(s) to substantiate this claim (not that I disagree with it).

**Extensibility**
- `otoole follows a strategy-based software design pattern` - do you mean `strategy design pattern`? Please provide a reference, so there is no doubt regarding what is meant.

**Publications**
- `pedagogical paper` - I think I understand what is meant, but this reads odd.

**Installation and Example**
- `otoole is deployed to the Python Packaging Index` - I believe this should say `published` instead of `deployed`.
