+++ paper/paper.bib
@@ -182,3 +182,13 @@
     title = {Calliope: a multi-scale energy systems modelling framework}, 
     journal = {Journal of Open Source Software} 
 } 
+
+@book{gamma1994,
+  title = {Design patterns},
+  author = {Gamma, Erich and Helm, Richard and Johnson, Ralph and Vlissides, John},
+  publisher = {Addison Wesley},
+  month = oct,
+  year = {1994},
+  address = {Boston, MA}
+  ISBN = {0201633612},
+}
