+++ paper/paper.bib
@@ -189,6 +189,6 @@
   publisher = {Addison Wesley},
   month = oct,
   year = {1994},
-  address = {Boston, MA}
+  address = {Boston, MA},
   ISBN = {0201633612},
 }
