+++ paper/paper.bib
@@ -140,22 +140,22 @@
 	author = {{Gurobi Optimization, LLC}},
 	year = {2022}
 }
-@article{groissbock2019, 
-    title = {Are open source energy system optimization tools mature enough for serious use?}, 
-    volume = {102}, 
-    issn = {1364-0321}, 
-    url = {https://www.sciencedirect.com/science/article/pii/S1364032118307743}, 
-    doi = {10.1016/j.rser.2018.11.020}, 
-    abstract = {Historically, energy system tools were predominantly proprietary and not shared with others. In recent years, there has been an increase in developing open source tools by international research and development organizations. More than half of the open energy modeling (openmod) initiative listed tools are based on the freely available scripting language Python. Previous comparisons of energy and power system modeling tools focused on comparisons such as which tool category (e.g. optimization, simulation) or energy demand (e.g. electricity, cooling, and heating) can be considered. Until now, the assessment of incorporated functions such as unit commitment (UC) or optimum power flow (OPF) has been ignored. Therefore, this work assesses 31 mostly open source tools based on 81 functions for their maturity. The result shows that available open source tools such as Switch, TEMOA, OSeMOSYS, and pyPSA are mature enough based on a function comparison with commercial or proprietary tools for serious use. Nevertheless, future commercial, as well as open source energy system analysis tools, have to consider more functions such as the impact of ambient air conditions and part-load behavior to allow better assessments of including high shares or renewable energy sources and other flexibility measures in existing and new energy systems.}, 
-    language = {en}, 
-    urldate = {2022-03-30}, 
-    journal = {Renewable and Sustainable Energy Reviews}, 
-    author = {Groissböck, Markus}, 
-    month = mar, 
-    year = {2019}, 
-    keywords = {Sector coupling, Optimization, Open source, Multi-energy system, Simulation}, 
-    pages = {234--248}, 
-} 
+@article{groissbock2019,
+    title = {Are open source energy system optimization tools mature enough for serious use?},
+    volume = {102},
+    issn = {1364-0321},
+    url = {https://www.sciencedirect.com/science/article/pii/S1364032118307743},
+    doi = {10.1016/j.rser.2018.11.020},
+    abstract = {Historically, energy system tools were predominantly proprietary and not shared with others. In recent years, there has been an increase in developing open source tools by international research and development organizations. More than half of the open energy modeling (openmod) initiative listed tools are based on the freely available scripting language Python. Previous comparisons of energy and power system modeling tools focused on comparisons such as which tool category (e.g. optimization, simulation) or energy demand (e.g. electricity, cooling, and heating) can be considered. Until now, the assessment of incorporated functions such as unit commitment (UC) or optimum power flow (OPF) has been ignored. Therefore, this work assesses 31 mostly open source tools based on 81 functions for their maturity. The result shows that available open source tools such as Switch, TEMOA, OSeMOSYS, and pyPSA are mature enough based on a function comparison with commercial or proprietary tools for serious use. Nevertheless, future commercial, as well as open source energy system analysis tools, have to consider more functions such as the impact of ambient air conditions and part-load behavior to allow better assessments of including high shares or renewable energy sources and other flexibility measures in existing and new energy systems.},
+    language = {en},
+    urldate = {2022-03-30},
+    journal = {Renewable and Sustainable Energy Reviews},
+    author = {Groissböck, Markus},
+    month = mar,
+    year = {2019},
[…]
