### 2023-09-27 bf43d3b
update footnote
