### 2023-09-27 #5511 @trevorb1
Hi @olejandro. Thank you so much for taking the time to review the paper! Please find below point-by-point responses to all of your comments. If any response is unclear, needs further explanation/discussion, or requires further attention, please let us know. Thank you!

**Currently it focuses solely on the OSeMOSYS users, however the problems that it is attempting to solve are common to other energy modelling communities. How does otoole, which is developed for OseMOSYS, differ from similar tools that were developed for other energy system models / model generators / frameworks?**

Good point! We have added a paragraph that explains the differences between how other open-source frameworks, such as PyPSA or Calliope, handle data, why it is different in OSeMOSYS’ case, and how otoole helps fill this need. Please see [this commit](https://github.com/OSeMOSYS/otoole/commit/845782f9ebb32fe1327a520eb02c9bd8f3dfe1f9), or specifically [this paragraph](https://github.com/OSeMOSYS/otoole/blob/f0d3a1c0839def6776a1c950f64b9cc269cf90c2/paper/paper.md?plain=1#L38), for the addition.

**`is a highly cited and reputable open-source framework` - please provide a reference(s) to substantiate this claim (not that I disagree with it).**

Good catch! We have added a reference that identifies OSeMOSYS as one of only a handful of open energy modeling frameworks mature enough to be used for energy planning studies. Moreover, we added a footnote about citations numbers via a SCOPUS search. Please see [this line](https://github.com/OSeMOSYS/otoole/blob/f0d3a1c0839def6776a1c950f64b9cc269cf90c2/paper/paper.md?plain=1#L34) for the reference update and [this line](https://github.com/OSeMOSYS/otoole/blob/f0d3a1c0839def6776a1c950f64b9cc269cf90c2/paper/paper.md?plain=1#L36) for the footnote.

**`otoole follows a strategy-based software design pattern` - do you mean strategy design pattern? Please provide a reference, so there is no doubt regarding what is meant.**

Yes, we mean strategy design pattern. We have updated the wording in the paper to reflect this and added a reference. Please see [this line](https://github.com/OSeMOSYS/otoole/blob/f0d3a1c0839def6776a1c950f64b9cc269cf90c2/paper/paper.md?plain=1#L54) for the update.

**`otoole is deployed to the Python Packaging Index` - I believe this should say `published` instead of `deployed`.**

Agreed. The wording has been updated. Please see [this line](https://github.com/OSeMOSYS/otoole/blob/f0d3a1c0839def6776a1c950f64b9cc269cf90c2/paper/paper.md?plain=1#L64) for the update.
