### 2023-10-11 #5511 @olejandro
@trevorb1 @fraukewiese apologies for the slow response!
Yes, I am generally satisfied with changes and answers so far. Thanks @trevorb1!
One last comment on the paper that I feel would strengthen it is to expand the comparison with other open source frameworks (i.e. addressing my first point) with:
- open source tool(s) that are intended to be model/framework agnostic, e.g. [ixmp](https://github.com/iiasa/ixmp4) / [pyam](https://github.com/IAMconsortium/pyam), [SPINE](https://github.com/spine-tools/Spine-Toolbox) or any other.
- example(s) of commercial solutions, e.g. the TIMES community is using one, and the main benefits of keeping things open source.

I'll come back with any comments on the software / repository towards the end of the week.
