### 2023-10-16 #5511 @olejandro
@trevorb1 @fraukewiese I have mostly completed the checklist.
I have experience a few issues with the software (I tried it on Windows):
- some of the [tests fail](https://github.com/openjournals/joss-reviews/files/12912789/tests.txt);
- an attempt to generate `res.png` results in `FileNotFoundError: [WinError 2] "dot" not found in path`. `res.graphml` does get generated;
- validation example leads to `AttributeError: 'NoneType' object has no attribute 'keys'` after validation starts.

Let me know if I should open an issue in the original repository for any of the above (or whether I am doing anything wrong).

One last point, it would be great if there was a URL to contributing guidelines included in the readme.
