### 2023-10-20 OSeMOSYS/otoole#202 @trevorb1
[Bug]: Tests fail on Windows 
### The Issue

As flagged by @olejandro in the JOSS review [here](https://github.com/openjournals/joss-reviews/issues/5511#issuecomment-1763801248), tests are failing when running on Windows.

See [here](https://github.com/OSeMOSYS/otoole/files/13057677/tests.txt) for a full log output

### Expected Behavior

I would expect all tests to pass

### Steps To Reproduce

1. Run `tox` from a Windows machine

### Log output

_No response_

### Operating System

Windows

### What version of otoole are you running?

1.1.0

### Possible Solution

_No response_

### Anything else?

_No response_
