### 2023-10-20 #5511 @trevorb1
Hi @olejandro, thank you for the review! I just confirmed on my side that the tests are failing on Windows as well. Thanks for flagging this! I have just created a new issue for that [here](https://github.com/OSeMOSYS/otoole/issues/202), as I am about to start working on it. Please feel free to add any information to the issue ticket, and/or create new issues for the other bugs you flagged. :)
