### 2023-10-21 d9a78ee
update param and set int dtypes to int64
