### 2023-10-21 019a096
fix int64 check
