### 2023-10-22 6066f34
fix int to int64 issue in read_strategies
