### 2023-10-23 ad87fd2
fix NamedTemporaryFile tests for Windows
