+++ tests/test_cli.py
@@ -41,15 +41,20 @@ class TestConvert:
         result = run(["otoole", "--version"], capture_output=True)
         assert result.stdout.strip().decode() == str(__version__)
 
+    def test_help(self):
+        commands = ["otoole", "-v", "convert", "--help"]
+        expected = "usage: otoole convert [-h]"
+        actual = run(commands, capture_output=True)
+        assert expected in str(actual.stdout)
+        assert actual.returncode == 0, print(actual.stdout)
+
     temp = mkdtemp()
-    temp_excel = NamedTemporaryFile(suffix=".xlsx")
-    temp_datafile = NamedTemporaryFile(suffix=".dat")
     simplicity = os.path.join("tests", "fixtures", "simplicity.txt")
     config_path = os.path.join("tests", "fixtures", "config.yaml")
 
     test_data = [
-        (["otoole", "-v", "convert", "--help"], "usage: otoole convert [-h]"),
         (
+            "excel",
             [
                 "otoole",
                 "-v",
@@ -57,12 +62,13 @@ class TestConvert:
                 "datafile",
                 "excel",
                 simplicity,
-                temp_excel.name,
+                "convert_to_file_path",  # replaced with NamedTemporaryFile
                 config_path,
             ],
             "",
         ),
         (
+            "datafile",
             [
                 "otoole",
                 "-v",
@@ -70,19 +76,34 @@ class TestConvert:
                 "datafile",
                 "datafile",
                 simplicity,
-                temp_datafile.name,
+                "convert_to_file_path",  # replaced with NamedTemporaryFile
                 config_path,
             ],
             "",
         ),
     ]
 
-    @mark.parametrize("commands,expected", test_data, ids=["help", "excel", "datafile"])
-    def test_convert_commands(self, commands, expected):
-        actual = run(commands, capture_output=True)
-        assert expected in str(actual.stdout)
-        print(" ".join(commands))
-        assert actual.returncode == 0, print(actual.stdout)
+    @mark.parametrize(
+        "convert_to,commands,expected", test_data, ids=["excel", "datafile"]
+    )
+    def test_convert_commands(self, convert_to, commands, expected):
+        if convert_to == "datafile":
+            temp = NamedTemporaryFile(suffix=".dat", delete=False, mode="w")
+        elif convert_to == "excel":
+            temp = NamedTemporaryFile(suffix=".xlsx", delete=False, mode="w")
+        else:
+            raise NotImplementedError
+        try:
+            commands_adjusted = [
+                x if x != "convert_to_file_path" else temp.name for x in commands
+            ]
+            actual = run(commands_adjusted, capture_output=True)
+            assert expected in str(actual.stdout)
+            print(" ".join(commands_adjusted))
+            assert actual.returncode == 0, print(actual.stdout)
+        finally:
+            temp.close()
+            os.unlink(temp.name)
 
     test_errors = [
         (
@@ -98,58 +119,67 @@ class TestConvert:
 
     def test_convert_datafile_datafile_no_user_config(self):
         simplicity = os.path.join("tests", "fixtures", "simplicity.txt")
-        temp_datafile = NamedTemporaryFile(suffix=".dat")
-        commands = [
-            "otoole",
-            "convert",
-            "datafile",
-            "datafile",
-            simplicity,
-            temp_datafile.name,
-        ]
-        actual = run(commands, capture_output=True)
-        assert actual.returncode == 2
+        temp_datafile = NamedTemporaryFile(suffix=".dat", delete=False, mode="w")
+        try:
+            commands = [
+                "otoole",
+                "convert",
+                "datafile",
+                "datafile",
+                simplicity,
+                temp_datafile.name,
+            ]
[…]
