### 2023-10-23 f84d1f7
add Windows to CI
