+++ docs/examples.rst
@@ -215,6 +215,20 @@ Run the following command, where the RES will be saved as the file ``res.png``::
 
     $ otoole viz res excel simplicity.xlsx res.png config.yaml
 
+.. WARNING::
+    If you encounter a ``graphviz`` dependency error, please install it in your
+    virtual environment through::
+
+        pip install graphviz # if using pip
+        conda install graphviz # if using conda
+
+    Alternatively, you can install it on your system via the commands::
+
+        sudo apt install graphviz # if on Ubuntu
+        brew install graphviz # if on Mac
+
+    Or from the graphviz_ website if on Windows.
+
 2. View the RES
 ~~~~~~~~~~~~~~~
 Open the newly created file, ``res.png`` and the following image should be
@@ -486,3 +500,4 @@ will also flag it as an isolated fuel. This means the fuel is unconnected from t
 .. _CPLEX: https://www.ibm.com/products/ilog-cplex-optimization-studio/cplex-optimizer
 .. _Anaconda: https://www.anaconda.com/
 .. _Gurobi: https://www.gurobi.com/
+.. _graphviz: https://www.graphviz.org/download/
