+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 amply>=0.1.4
 docutils<0.18
+graphviz
 flatten_dict
 Jinja2<3.1
 networkx
