+++ setup.cfg
@@ -53,6 +53,7 @@ install_requires =
     flatten_dict
     openpyxl
     pydantic>=2
+    graphviz
 [options.packages.find]
 where = src
 exclude =
