### 2023-11-03 d7f668e
fix graphviz dependency
