+++ paper/paper.bib
@@ -193,24 +193,24 @@
   ISBN = {0201633612},
 }
 
-@software{huppmann2021, 
-    author = {Huppmann, Daniel and Gidden, Matthew J. and Nicholls, Zebedee and Hörsch, Jonas and Lamboll, Robin D. and Kishimoto, Paul Natsuo and Burandt, Thorsten and Fricko, Oliver and Byers, Edward and Kikstra, Jarmo S. and Brinkerink, Maarten and Budzinski, Maik and Maczek, Florian and Zwickl-Bernhard, Sebastian and Welder, Lara and Alvarez Quispe, Erik Francisco and Smith, Christopher J.}, 
-    doi = {10.12688/openreseurope.13633.2}, 
+@software{huppmann2021,
+    author = {Huppmann, Daniel and Gidden, Matthew J. and Nicholls, Zebedee and Hörsch, Jonas and Lamboll, Robin D. and Kishimoto, Paul Natsuo and Burandt, Thorsten and Fricko, Oliver and Byers, Edward and Kikstra, Jarmo S. and Brinkerink, Maarten and Budzinski, Maik and Maczek, Florian and Zwickl-Bernhard, Sebastian and Welder, Lara and Alvarez Quispe, Erik Francisco and Smith, Christopher J.},
+    doi = {10.12688/openreseurope.13633.2},
     volume={1},
     journal={Open Research Europe}
-    title = {pyam: analysis and visualization of integrated-assessment and macro-energy scenarios [version 2; peer review: 3 approved]}, 
+    title = {pyam: analysis and visualization of integrated-assessment and macro-energy scenarios [version 2; peer review: 3 approved]},
 }
 
-@article{kiviluoma2021, 
-    title = {Spine Toolbox: A flexible open-source workflow management system with scenario and data management}, 
-    journal = {SoftwareX}, 
-    volume = {17}, 
-    pages = {100967}, 
-    year = {2022}, 
-    issn = {2352-7110}, 
-    doi = {https://doi.org/10.1016/j.softx.2021.100967}, 
-    url = {https://www.sciencedirect.com/science/article/pii/S2352711021001886}, 
-    author = {Juha Kiviluoma and Fabiano Pallonetto and Manuel Marin and Pekka T. Savolainen and Antti Soininen and Per Vennström and Erkka Rinne and Jiangyi Huang and Iasonas Kouveliotis-Lysikatos and Maren Ihlemann and Erik Delarue and Ciara O’Dwyer and Terence O’Donnel and Mikael Amelin and Lennart Söder and Joseph Dillon}, 
-    keywords = {Workflow management, Data management, Modelling, Optimisation, Open source, Scenario}, 
-    abstract = {The Spine Toolbox is open-source software for defining, managing, simulating and optimising energy system models. It gives the user the ability to collect, create, organise, and validate model input data, execute a model with selected data and finally archive and visualise results/output data. Spine Toolbox has been designed and developed to support the creation and execution of multivector energy integration models. It conveniently facilitates the linking of models with different scopes, or spatio-temporal resolutions, through the user interface. The models can be organised as a direct acyclic graph and efficiently executed through the embedded workflow management engine. The software helps users to import and manage data, define models and scenarios and orchestrate projects. It supports a self-contained and shareable entity-relationship data structure for storing model parameter values and the associated data. The software is developed using the latest Python environment and supports the execution of plugins. It is shipped in an installation package as a desktop application for different operating systems.} 
-} 
+@article{kiviluoma2021,
+    title = {Spine Toolbox: A flexible open-source workflow management system with scenario and data management},
+    journal = {SoftwareX},
+    volume = {17},
+    pages = {100967},
[…]
