+++ paper/paper.bib
@@ -197,7 +197,7 @@
     author = {Huppmann, Daniel and Gidden, Matthew J. and Nicholls, Zebedee and Hörsch, Jonas and Lamboll, Robin D. and Kishimoto, Paul Natsuo and Burandt, Thorsten and Fricko, Oliver and Byers, Edward and Kikstra, Jarmo S. and Brinkerink, Maarten and Budzinski, Maik and Maczek, Florian and Zwickl-Bernhard, Sebastian and Welder, Lara and Alvarez Quispe, Erik Francisco and Smith, Christopher J.},
     doi = {10.12688/openreseurope.13633.2},
     volume={1},
-    journal={Open Research Europe}
+    journal={Open Research Europe},
     title = {pyam: analysis and visualization of integrated-assessment and macro-energy scenarios [version 2; peer review: 3 approved]},
 }
 
