+++ paper/paper.bib
@@ -208,7 +208,7 @@
     pages = {100967},
     year = {2022},
     issn = {2352-7110},
-    doi = {https://doi.org/10.1016/j.softx.2021.100967},
+    doi = {10.1016/j.softx.2021.100967},
     url = {https://www.sciencedirect.com/science/article/pii/S2352711021001886},
     author = {Juha Kiviluoma and Fabiano Pallonetto and Manuel Marin and Pekka T. Savolainen and Antti Soininen and Per Vennström and Erkka Rinne and Jiangyi Huang and Iasonas Kouveliotis-Lysikatos and Maren Ihlemann and Erik Delarue and Ciara O’Dwyer and Terence O’Donnel and Mikael Amelin and Lennart Söder and Joseph Dillon},
     keywords = {Workflow management, Data management, Modelling, Optimisation, Open source, Scenario},
