### 2023-11-09 #5511 @trevorb1
Thank you again for your feedback, @olejandro! Please find below point by point responses to your comments. If anything is unclear or needs further explanation, please let me know.

 **One last comment on the paper that I feel would strengthen it is to expand the comparison with other open source frameworks**
Apologies for the confusion on your original comment. Please find an update in [this commit](https://github.com/OSeMOSYS/otoole/commit/7872c611ce304132ed031d0af8b49065fb70d78b) that states how `otoole` fits into the literature.

**Some of the [Windows] tests fail**
Good catch! Windows tests have been updated in [this PR](https://github.com/OSeMOSYS/otoole/pull/203). Additionally, Windows tests are now included in the CI.

**An attempt to generate res.png results in FileNotFoundError: [WinError 2] "dot" not found in path. res.graphml does get generated;**
This seems to be a `graphviz` dependency issue, where it must be installed on the system. We have added instructions on how to do this for both Windows and UNIX in [this PR](https://github.com/OSeMOSYS/otoole/pull/205). Please see [here](https://otoole.readthedocs.io/en/latest/examples.html#otoole-visualise) for the documentation updates.

**Validation example leads to AttributeError: 'NoneType' object has no attribute 'keys' after validation starts.**
Unfortunately, I was not able to recreate this issue. I tried on both Windows and Linux, and the validation worked. On Windows, this is the command I used using otoole version `1.1.0`.

```bash
otoole validate datafile data.txt config.yaml --validate_config validate.yaml
```

If you are still experiencing an issue, maybe it will make sense to move this to its own issue ticket. In the issue ticket, can you please:
1. Confirm the version of otoole you are running through the `otoole --version` command
2. Please run the validation command in verbose mode (ie. `otoole -v validate ...`) to print out the traceback error.

**it would be great if there was a URL to contributing guidelines included in the readme.**
Agreed! This has been updated in [this PR](https://github.com/OSeMOSYS/otoole/pull/206); additionally, a wider selection of badges have been added to the readme.
