### 2023-11-22 #5511 @fraukewiese
@trevorb1 : Thanks for the updates :)
@olejandro : Are you satisfied with the changes and replies to your comments/raised issues? And do you still experience this issue: "Validation example leads to AttributeError: 'NoneType' object has no attribute 'keys' after validation starts." (see @trevorb1 's explanation above.
