### 2023-12-05 #5511 @fraukewiese
@trevorb1 : There are a few minor issues on the paper:
- line 38: add a space between pyam and (Huppmann ...)
- line 39: add a space between Spine-Toolbox and (Kiviluoma ...)
- line 75/76: ... by the user, the ... instead of a fullstop after user
- line 92: should it be "such as the Climate ..." instead of "such of the Climate..."?
- Please add a date to the source Huppmann et al.
