+++ paper/paper.bib
@@ -199,6 +199,7 @@
     volume={1},
     journal={Open Research Europe},
     title = {pyam: analysis and visualization of integrated-assessment and macro-energy scenarios [version 2; peer review: 3 approved]},
+    year = {2021},
 }
 
 @article{kiviluoma2021,
