### 2023-12-07 #5511 @trevorb1
Hi @fraukewiese! Thank you for your final edits! All points have been addressed in [this commit](https://github.com/OSeMOSYS/otoole/commit/b5ab2880a295510e9fc261c73087a7c74dfeeb27). On your comment for line 75/76, I instead did a slight reword  just to make the sentence more clear, instead of adding a comma. Please let me know if the line is still unclear and I will update. Thank you again!
