+++ .github/workflows/python.yaml
@@ -28,8 +28,6 @@ jobs:
         run: |
           tox
       - name: Upload coverage data to converalls.io
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          pip install --upgrade coveralls
-          coveralls --service=github
+        uses: coverallsapp/github-action@v2
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
