+++ .github/workflows/python.yaml
@@ -31,3 +31,4 @@ jobs:
         uses: coverallsapp/github-action@v2
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
+          debug: True
