+++ .github/workflows/python.yaml
@@ -28,7 +28,8 @@ jobs:
         run: |
           tox
       - name: Upload coverage data to converalls.io
-        uses: coverallsapp/github-action@v2
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          debug: True
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        run: |
+          pip install --upgrade coveralls
+          coveralls --service=github
