+++ .github/workflows/python.yaml
@@ -30,6 +30,7 @@ jobs:
       - name: Upload coverage data to converalls.io
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        if: ${{ matrix.os }} == "ubuntu-latest" && ${{ matrix.python-version }} == "ubuntu-latest"
         run: |
           pip install --upgrade coveralls
           coveralls --service=github
