### 2023-12-11 5db9938
update syntax in action
