### 2023-12-12 #5511 @trevorb1
Hi @fraukewiese! Please find attached all final information requested:
- **Software Version**: `otoole` version `v1.1.2` has been released today. See release version on [PyPI](https://pypi.org/project/otoole/) and on [GitHub](https://github.com/OSeMOSYS/otoole/releases/tag/v1.1.2)
- **Software Archive**: Please see `otoole` version `v1.1.2` archived on Zenodo [here](https://zenodo.org/records/10360538)
  - **Title**: "otoole: OSeMOSYS Tools for Energy Work" matches the paper title
  - **Authors**: Trevor Barnes and Will Usher. Same order as paper
  - **DOI**: `otoole` version `v1.1.2` Zenodo DOI is [10.5281/zenodo.10360538](https://zenodo.org/records/10360538)

If you need any other information from my end, please just let me know. Thank you so much in advance! 😊
