### 2023-12-15 3a346de
Modified citation file with template for JOSS article
