### 2023-12-15 cc90480
Update coveralls workflow to use lcov
