+++ setup.cfg
@@ -87,7 +87,7 @@ console_scripts =
 # CAUTION: --cov flags may prohibit setting breakpoints while debugging.
 #          Comment those flags to avoid this pytest issue.
 addopts =
-    --cov otoole --cov-report html
+    --cov otoole --cov-report lcov
     --verbose
     -s
     # --log-cli-level=10
