### 2023-12-15 dcd5a49
Regression in tests caused by pandas fixed by pandas 2.1.4
