+++ paper/paper.bib
@@ -115,7 +115,7 @@
   pages = {696--716},
   author = {Eunice Pereira Ramos and Vignesh Sridharan and Thomas Alfstad and Taco Niet and Abhishek Shivakumar and Mark Idwal Howells and Holger Rogner and Francesco Gardumi},
   title = {Climate,  Land,  Energy and Water systems interactions {\textendash} From key concepts to model implementation with {OSeMOSYS}},
-  journal = {Environmental Science {\&}amp$\mathsemicolon$ Policy}
+  journal = {Environmental Science & Policy}
 }
 
 @misc{GNULinearProgramming2012,
