### 2023-12-19 #5511 @trevorb1
@fraukewiese I have addressed the `\mathsemicolon` issue with [this commit](https://github.com/OSeMOSYS/otoole/commit/0a239c9f536ff532a79e7739afde9ebc237fe241). That was my mistake and there was a formatting issue with a Journal's name.
